chore: mover formulario a su propio componente y agregar etiquetas de obligatorio
